+++ extensions/akkoma/src/components/VisibilityDropdown.tsx
@@ -13,7 +13,7 @@ const VisibilityDropdown = () => {
   const { defaultVisibility }: Preferences = getPreferenceValues();
 
   return (
-    <Form.Dropdown id="visibility" title="Visibility" storeValue={true} defaultValue={defaultVisibility}>
+    <Form.Dropdown id="visibility" title="Visibility" storeValue={true} defaultValue={defaultVisibility || undefined}>
       {visibilityOptions.map(({ value, title, icon }) => (
         <Form.Dropdown.Item
           key={value}