+++ extensions/akkoma/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Update] - 2023-05-11
   - Update Raycast API to 1.51.0
+  - Refactor Request API
 ## [Initial Version] - 2023-04-06